refactor(frontend): consolidate BROWSER_SOURCE_MODES into single source
Merge the two name-colliding BROWSER_SOURCE_MODES constants (string array
in graphUtils.ts, rich object array in sourceModes.ts) so menu metadata
and the id list used by isBrowserSourceMode cannot drift.
